requirements.txt
fastapi
strawberry-graphql[fastapi]
uvicorn
uvicorn[standard]
pymongo
